WIP refactor data block
