+++ src/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(c_sources
     ${CMAKE_CURRENT_SOURCE_DIR}/rex-header.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/rex-data-block.c
     ${CMAKE_CURRENT_SOURCE_DIR}/core.c
     ${CMAKE_CURRENT_SOURCE_DIR}/list.c
     ${CMAKE_CURRENT_SOURCE_DIR}/util.c
@@ -8,6 +9,7 @@ set(c_sources
 set(c_headers
     ${CMAKE_CURRENT_SOURCE_DIR}/config.h
     ${CMAKE_CURRENT_SOURCE_DIR}/rex-header.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/rex-data-block.h
     ${CMAKE_CURRENT_SOURCE_DIR}/core.h
     ${CMAKE_CURRENT_SOURCE_DIR}/global.h
     ${CMAKE_CURRENT_SOURCE_DIR}/linmath.h
