+++ src/rex-data-block.h
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2018 Robotic Eyes GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.*
+ */
+#pragma once
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct rex_block_header
+{
+    uint16_t type;
+    uint16_t version;
+    uint32_t sz;  // data block size w/o header
+    uint64_t id;
+};
+
+/*
+ * Reads a data block header from the current file position
+ */
+int rex_block_header_read (FILE *fp, struct rex_block_header *header);
+
+/*
+ * Reads the data block from the current file position with the given size len
+ */
+int rex_block_read (FILE *fp, uint8_t *block, uint32_t len);
+
+#ifdef __cplusplus
+}
+#endif
+
