+++ viewer/rex-view.c
@@ -25,6 +25,7 @@
 #include "global.h"
 #include "mesh.h"
 #include "mesh_group.h"
+#include "rex-data-block.h"
 #include "shader.h"
 #include "status.h"
 #include "util.h"
@@ -261,7 +262,7 @@ int loadrex (const char *file)
     for (int i = 0; i < header.nr_datablocks; i++)
     {
         struct rex_block_header block_header;
-        rex_read_data_block_header (fp, &block_header);
+        rex_block_header_read(fp, &block_header);
 
         if (block_header.type == Mesh)
         {
